requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
flask-migrate==4.0.7
pymysql==1.1.1
cryptography==42.0.8
python-dotenv==1.0.1
croniter
werkzeug==3.0.3
email-validator